builder: Add option to skip os-core-update-index generation
os-core-update-index was decided to be removed in the future, but
we need to stop creating that for 3rd party content. Add a build time
flag to skip index creation. If flag is used index won't be processed
and previous index will be carried out if it exists. So build won't break
and no files will be left behind.

If --skip-update-index is used since first build, no index will be ever built.

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>